# public/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Blacksport",
        "version": "1.0.0"
    },
    "paths": {
        "/api/about/{locale?}": {
            "get": {
                "tags": [
                    "about"
                ],
                "summary": "About Page",
                "description": "Getting Ambassadors, Partners, Texts",
                "operationId": "App\\Http\\Controllers\\AboutController::index",
                "parameters": [
                    {
                        "name": "Locale",
                        "in": "path",
                        "description": "Language",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "ru, en, uk"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "page": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "name": {
                                                        "type": "string"
                                                    },
                                                    "page_key": {
                                                        "type": "string"
                                                    },
                                                    "meta_description": {
                                                        "type": "string"
                                                    },
                                                    "noindex": {
                                                        "type": "integer"
                                                    },
                                                    "nofollow": {
                                                        "type": "integer"
                                                    },
                                                    "created_at": {
                                                        "type": "string"
                                                    },
                                                    "updated_at": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "ambassadors": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "description": {
                                                        "type": "string"
                                                    },
                                                    "image": {
                                                        "type": "string"
                                                    },
                                                    "created_at": {
                                                        "type": "string"
                                                    },
                                                    "updated_at": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "partners": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "description": {
                                                        "type": "string"
                                                    },
                                                    "image": {
                                                        "type": "string"
                                                    },
                                                    "created_at": {
                                                        "type": "string"
                                                    },
                                                    "updated_at": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "achievements": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "title": {
                                                        "type": "string"
                                                    },
                                                    "description": {
                                                        "type": "string"
                                                    },
                                                    "created_at": {
                                                        "type": "string"
                                                    },
                                                    "updated_at": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/partners/{locale?}": {
            "get": {
                "tags": [
                    "partner"
                ],
                "summary": "List of partners",
                "description": "Getting partners",
                "operationId": "App\\Http\\Controllers\\PartnerController::index",
                "parameters": [
                    {
                        "name": "Locale",
                        "in": "path",
                        "description": "Language",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "ru, en, uk"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "partners": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "description": {
                                                        "type": "string"
                                                    },
                                                    "image": {
                                                        "type": "string"
                                                    },
                                                    "image_type": {
                                                        "type": "string"
                                                    },
                                                    "created_at": {
                                                        "type": "string"
                                                    },
                                                    "updated_at": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "security": [
        []
    ]
}